+++ src/builder/mod.rs
@@ -40,8 +40,15 @@ pub(crate) fn build_file(pair: Pair<Rule>) -> Result<File> {
                         let alias_pair = inner.next().unwrap();
                         let alias = alias_pair.as_str().to_string();
 
+                        let components = if let Some(block_pair) = inner.next() {
+                            let list_pair = block_pair.into_inner().next().unwrap();
+                            Some(list_pair.into_inner().map(|p| p.as_str().to_string()).collect())
+                        } else {
+                            None
+                        };
+
                         TopLevelItem::Import(ImportDef {
-                            def: ImportDefType::Module(path, alias),
+                            def: ImportDefType::Module(path, alias, components),
                             span: Span::from_pair(&child),
                         })
                     }