+++ src/error.rs
@@ -193,6 +193,8 @@ fn rule_to_human(rule: Rule) -> &'static str {
         Rule::prop_value => "a property value",
 
         // === Imports ===
+        Rule::layout_list => "a layout list",
+        Rule::import_layout_block => "am import layout block",
         Rule::import_stmt => "an import statement",
         Rule::import_lib_stmt => "a library import statement",